requirements.txt
fastmcp==2.11.3
mcp==1.13.0
openai==1.98.0
python-dotenv==1.1.1
PyQt5==5.15.11
qasync==0.27.1
#langgraph==0.2.34
#langchain
#langchain-openai
#langchain-core==0.3.6
google-adk==1.13.0
litellm==1.72.0
#litellm[proxy]
a2a-sdk==0.3.3
click==8.2.1